[metasrv] refactor: move raft test supporting utils into metasrv/raft/testing.rs
